Split annotations events code into new module
